Fix test_samples field to return integer values
- Convert samples_count to int to ensure test_samples is an integer
- This fixes the issue where test_samples was showing decimal values
- Now test_samples will correctly show total hours (e.g., 2400, 2376)

+++ train/train_dl.py
@@ -280,7 +280,7 @@ def train_dl_model(
         'param_count': count_parameters(model),
         'train_time_sec': round(time.time() - training_start_time, 2),
         'inference_time_sec': total_inference_time,
-        'samples_count': len(p_inv_matrix) * p_inv_matrix.shape[1],  # Number of test samples (total hours)
+        'samples_count': int(len(p_inv_matrix) * p_inv_matrix.shape[1]),  # Number of test samples (total hours)
         'predictions': final_preds_24h,  # 1D array of 24h-ahead predictions (for CSV)
         'y_true': final_gt_24h,          # 1D array of corresponding ground truth (for CSV)
         'predictions_all': p_inv_matrix,  # Full multi-step predictions (for potential analysis)
